refactor: place input content refactored
